Add ufbx_find_prop_concat()
This function is used to find properties with concatenated names. There was a buffer overflow
bug earlier in ufbxi_update_constraint() as node names could actually be over 256 characters.
Unfortunately we can't allocate memory in that function so we need to be able to search for
concatenated property names without allocating memory.